requirements.txt
numpy==1.23.4
Pillow==9.3.0
playsound==1.3.0
tk==0.1.0

sounddevice~=0.4.5
soundfile~=0.11.0
pycaw~=20220416